Updated short report template
Fixed an issue in which the document failed to render. Not too sure if this issue was due to a slightly different environment required to properly run the files. Changes made also change the way in which certain parameters are named such as email and report title. Couldn't get file to render references as needed from references.bib yet.